Bchron: Radiocarbon dating, age-depth modelling, relative sea level rate estimation, and non-parametric phase modelling

This package enables quick calibration of radiocarbon dates under various calibration curves (including user generated ones); Age-depth modelling as per the algorithm of Haslett and Parnell (2008); Relative sea level rate estimation incorporating time uncertainty in polynomial regression models; and non-parametric phase modelling via Gaussian mixtures as a means to determine the activity of a site (and as an alternative to the Oxcal function SUM)

Bchron functions

The most important functions are BchronCalibrate to calibrate radiocarbon (and non-radiocarbon) dates, Bchronology for the age-depth model of Haslett and Parnell (2008), BchronRSL to get rate estimates for relative sea level data, BchronDensity and BchronDensityFast for non-parametric phase modelling of age data. See the help files for these functions for examples. See the vignette for more complete documentation