Example chronology file for use with the BchronRSL function.

Some example chronology data for use with the BchronRSL function

data(TestChronData)

Format

A data frame with 27 observations on the following 6 variables:

id

ID names

ages

Ages in years BP

ageSds

Ages standard deviations in years BP

position

Depths in cm

thickness

Thicknesses in cm

calCurves

Calibration curve for each age

Source

Andrew C. Parnell and W. Roland Gehrels (2013) 'Using chronological models in late holocene sea level reconstructions from salt marsh sediments' In: I. Shennan, B.P. Horton, and A.J. Long (eds). Handbook of Sea Level Research. Chichester: Wiley