Plot calibrated dates from a BchronCalibrate run

Plots calibrated radiocarbon dates from a BchronCalibrate run. Has options to plot on a position (usually depth) scale if supplied with the original run

# S3 method for BchronCalibratedDates
plot(
  x,
  withPositions = ifelse(length(x) > 1 & !is.null(x[[1]]$positions), TRUE, FALSE),
  dateHeight = 100,
  dateLabels = TRUE,
  fillCol = "darkslategray",
  withHDR = TRUE,
  ageScale = c("bp", "bc", "b2k"),
  scaleReverse = TRUE,
  ...
)

Arguments

x
Output from BchronCalibrate

withPositions
Whether to plot with positions (i.e. using the position values as the y axis). By default TRUE if x has more than one date and contains positions

dateHeight
The height of the dates in the plot in the same units as the position values. Only relevant if withPositions=TRUE.

dateLabels
Whether to add the names of the dates to the left of them. Default TRUE

fillCol
A colour to fill the date densities when withPositions is TRUE, or HDR ranges when it is FALSE

withHDR
Whether to plot the 95% highest density region values

ageScale
Either bp for years before present, bc for years BC/AD (BC will be negative), b2k for years before 2000. Others not supported (yet).

scaleReverse
Whether to reverse the x-axis scale. Defaults to TRUE which works best for dates presented in e.g. years BP

...
Other arguments to plot (currently ignored)

Details

These plots are intended to be pretty basic and used simply for quick information. Users are encouraged to learn the R plotting features to produce publication quality graphics

See also