Predict ages of other positions for a BchronologyRun object

This function will predict the ages of new positions (usually depths) based on a previous run of the function Bchronology. It will also allow for thickness uncertainties to be included in the resulting ages, for example when the age of a particular event is desired

# S3 method for class 'BchronologyRun'
predict(
  object,
  newPositions,
  newPositionThicknesses = NULL,
  maxExtrap = 500,
  ...
)

Arguments

object

Output from a run of Bchronology

newPositions

A vector of new positions at which to find ages

newPositionThicknesses

A vector of thicknesses for the above positions. Must be the same length as newPositions. If NULL then assumed zero

maxExtrap

The maximum new of extrapolation attempts. It might be worth increasing this if you are extrapolating a long way from the other dated positions

...

Other arguments to predict (not currently supported)

Value

A matrix of dimension num_samples by num_positions so that each row represents a set of monotonic sample predicted ages

See also