Get sample ages from a set of Bchron calibrated dates

A function for extracting sample ages from Bchron calibrated dates

sampleAges(CalDates, n_samp = 10000)

Arguments

CalDates
A list created from either BchronCalibrate.

n_samp
The desired number of samples

Value

A vector of length n_samp containing sample ages for the specified date

Details

Sometimes it is useful to have a set of sample calendar ages for your calibrated dates. For example the samples might be required to create a credible/confidence interval, or to create another non-trivial function of calibrated dates, such as differences. By default the BchronCalibrate function provides a grid of ages and an associated density, similar to OxCal. This function extracts that information and uses the sample function to output the desired number of samples

See also

Examples

# Calibrate multiple ages and summarise them
ages = BchronCalibrate(ages=c(3445,11553,7456),ageSds=c(50,230,110),
                       calCurves=c('intcal13','intcal13','shcal13'))
# Get samples
age_samples = sampleAges(ages)
# Create a credible interval and the median for each date
apply(age_samples, 2, quantile, probs = c(0.05, 0.5, 0.95))

#>     Date1    Date2 Date3
#> 5%   3603 13053.00  8035
#> 50%  3709 13396.50  8228
#> 95%  3824 13794.05  8382