Plot calibrated dates from a BchronCalibrate run

Plots calibrated radiocarbon dates from a BchronCalibrate run. Has options to plot on a position (usually depth) scale if supplied with the original run

# S3 method for BchronCalibratedDates
plot(
  x,
  date = NULL,
  withPositions = ifelse(length(x) > 1 & !is.null(x[[1]]$positions) & !includeCal,
    TRUE, FALSE),
  includeCal = FALSE,
  dateHeight = 100,
  dateLabels = TRUE,
  dateLabelSize = 2,
  nudgeX = 0,
  nudgeY = 0,
  fillCol = rgb(47/255, 79/255, 79/255, 0.5),
  withHDR = TRUE,
  ageScale = c("bp", "bc", "b2k"),
  scaleReverse = TRUE,
  pathToCalCurves = system.file("data", package = "Bchron"),
  ...
)

Arguments

x
Output from BchronCalibrate

date
Either numbers or date names to plot (only used if multiple dates have been calibrated)

withPositions
Whether to plot with positions (i.e. using the position values as the y axis). By default TRUE if x has more than one date and contains positions

includeCal
Whether to plot the date alongside the calibration curve (with 95% uncertainty bands) and the normally distributed uncalibrated date.

dateHeight
The height of the dates in the plot in the same units as the position values. Only relevant if withPositions=TRUE.

dateLabels
Whether to add the names of the dates to the left of them. Default TRUE

dateLabelSize
Size of the date labels

nudgeX
The amount to move the date labels in the x direction. Can be negative. See geom_text for details

nudgeY
The amount to move the date labels in the y direction. Can be negative. See geom_text for details

fillCol
A colour to fill the date densities when withPositions is TRUE, or HDR ranges when it is FALSE

withHDR
Whether to plot the 95% highest density region values

ageScale
Either bp for years before present, bc for years BC/AD (BC will be negative), b2k for years before 2000. Others not supported (yet).

scaleReverse
Whether to reverse the x-axis scale. Defaults to TRUE which works best for dates presented in e.g. years BP

pathToCalCurves
The Bchron path to calibration curves. Defaults to the package location might need to be set to another folder if user defined calibration curves are being used

...
Other arguments to plot (currently ignored)

Details

These plots are intended to be pretty basic and used simply for quick information. Users are encouraged to learn the R plotting features to produce publication quality graphics

See also