Plot output from Bchronology

Plots output from a run of Bchronology

# S3 method for BchronologyRun
plot(
  x,
  dateHeight = 100,
  dateLabels = TRUE,
  dateLabelSize = 2,
  dateCol = rgb(47/255, 79/255, 79/255, 0.5),
  chronCol = "deepskyblue4",
  chronTransparency = 0.75,
  alpha = 0.95,
  nudgeX = 0,
  nudgeY = 0,
  expandX = if (dateLabels) {     c(0.1, 0) } else {     c(0, 0) },
  expandY = c(0.05, 0),
  ageScale = c("bp", "bc", "b2k"),
  scaleReverse = TRUE,
  ...
)

Arguments

x
The object created by Bchronology

dateHeight
The height of the dates in the plot (on the same scale as the positions)

dateLabels
Whether to label the dates on the vertical axis (default TRUE)

dateLabelSize
The size of the date labels

dateCol
The colour of the date labels

chronCol
The colour of the chronology uncertainty ribbon to be plotted

chronTransparency
The amount of transparency for the chronology ribbon

alpha
The credible interval of the chronology run to be plotted. Defaults to 95 percent

nudgeX
The amount to move the date labels in the x direction. Can be negative. See geom_text for details

nudgeY
The amount to move the date labels in the y direction. Can be negative. See geom_text for details

expandX
The amount to expand the horizontal axis in case part are missed off the plot. See expand_limits for details

expandY
The amount to expand the vertical axis in case part are missed off the plot. See expand_limits for details

ageScale
Either bp for years before present, bc for years BC/AD (BC will be negative), b2k for years before 2000. Others not supported (yet).

scaleReverse
Whether to reverse the x-axis scale. Defaults to TRUE which works best for dates presented in e.g. years BP

...
Other arguments to plot (currently ignored)

Details

Creates a simple plot of the chronology output. The height of the date densities in the plots can be manipulated via the dateHeight argument which is represented in the same units as the positions/depths provided. More detailed plots can be created by manipulating the Bchronology object as required.

See also