Summarise a Bchronology object

Summarise a Bchronology object

# S3 method for class 'BchronologyRun'
summary(
  object,
  type = c("quantiles", "outliers", "convergence", "sed_rate", "acc_rate", "max_var"),
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  acc_probs = c(0.05, 0.95),
  useExisting = TRUE,
  numPos = 3,
  ...,
  digits = max(3, getOption("digits") - 3)
)

Arguments

object

Output from a run of Bchronology

type

Type of output required. The default (quantiles) gives the quantiles of the ages for each position in predictPositions from Bchronology. The other options provide outlier probabilities, convergence diagnostics, accumulation rates, sedimentation rate, and positions of maximum age variance

probs

Probabilities (between 0 and 1) at which to summarise the predicted chronologies

acc_probs

The age range over which to calculate accumulation rates (the default is between the 5% and 95% quantiles of all predicted ages)

useExisting

Whether to use the predicted chronologies/positions to calculate the sedimentation rate (if TRUE - default) or to re-create them based on a unit-scaled position grid (if FALSE). The latter will be a little bit slower but will provide better sedimentation rate estimates if the original positions are not on a unit scale (e.g. each cm)

numPos

The number of positions at which to provide the maximum variance

...

Other arguments (not currently supported)

digits

Number of digits to report values

Value

A data frame summarising the Bchronology object depending on the type of summary requested. For quantiles, a data frame with the quantiles for each position. For outliers, a data frame with outlier probabilities. For convergence, a data frame with p-values for convergence diagnostics. For sedimentation and accumulation rates, a data frame with the rates at each position/age. For maximum variance, a vector of positions with maximum age variance.

See also